apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings
sqlalchemy>=2.0.25
alembic
psycopg[binary]>=3.1.18
langchain>=0.1.0
langgraph>=0.0.10
python-dotenv
httpx
playwright
openai
anthropic
celery[redis]
redis
pdfminer.six
python-multipart
pgvector
reportlab
python-jose[cryptography]
